implement login

+++ app/views.py
@@ -1,24 +1,16 @@
-from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth import authenticate, login
+from django.shortcuts import get_object_or_404, redirect, render
+
 from .models import Product, User
 
 
 def index(request):
-    return render(request, "index.html")
+    return render(request, "index.html", {"user": request.user})
 
 
-def login(request):
-    if request.POST:
-        email = request.POST.get("email")
-        password = request.POST.get("password")
-
-        user = User.objects.get(email=email, password=password)
-        if user:
-            # request.session
-            user = User.objects.get(email=email)
-            return redirect(f"/profile/{user.username}")
-
-    return render(request, "login.html")
+def about(request):
+    return render(request, "about.html", {"user": request.user})
 
 
 def register(request):
@@ -29,38 +21,65 @@ def register(request):
 
         if email and username and password:
             if User.objects.filter(email=email).exists():
-                messages.error(request, "A user with this email already exists.")
+                messages.error(request, "A user with this email already exists")
+                return redirect("/register")
             elif User.objects.filter(username=username).exists():
-                messages.error(request, "A user with this username already exists.")
+                messages.error(request, "A user with this username already exists")
+                return redirect("/register")
             else:
-                user = User(username=username, email=email, password=password)
-                user.save()
-                messages.success(request, "Registration successful!")
+                User.objects.create_user(
+                    username=username, email=email, password=password
+                )
+                messages.success(request, "Registration successful")
                 return redirect("/login")
 
-    return render(request, "register.html")
+    return render(request, "register.html", {"user": request.user})
 
 
-def profile(request, username):
-    user = get_object_or_404(User, username=username)
-    return render(request, "profile.html", {"user": user})
+def login_user(request):
+    if request.POST:
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+
+        user = authenticate(username=username, password=password)
+        if user:
+            login(request, user)
+            return redirect("/profile")
+        else:
+            messages.error(request, "Username or password is incorrect")
+            return redirect("/login")
+
+    return render(request, "login.html", {"user": request.user})
 
 
-def about(request):
-    return render(request, "about.html")
+def profile(request, username=None):
+    if username:
+        try:
+            user = User.objects.get(username=username)
+            return render(request, "profile.html", {"user": user})
+        except User.DoesNotExist:
+            messages.error(request, "User does not exist")
+            return redirect("/login")  # should go to 404
 
-
-def market(request):
-    products = Product.objects.all()
-    return render(request, "market.html", {"products": products})
+    if request.user.is_authenticated:
+        return render(request, "profile.html", {"user": request.user})
+    return redirect("/login")
 
 
 def product(request, id):
-    product = Product.objects.get(pk=id)
-    return render(request, "product.html", {"product": product})
+    try:
+        product = Product.objects.get(pk=id)
+    except Product.DoesNotExist:
+        messages.error(request, "product does not exist")
+        return redirect("/market")  # should go to 404
+
+    return render(request, "product.html", {"user": request.user, "product": product})
 
 
 def add(request):
+    if not request.user.is_authenticated:
+        return redirect("/market")
+
     if request.POST:
         name = request.POST.get("product-name")
         price = request.POST.get("price")
@@ -69,11 +88,16 @@ def add(request):
             product = Product(name=name, price=price, desc=desc)
             product.save()
             return redirect(f"/product/{product.id}")
-    return render(request, "add.html")
+
+    return render(request, "add.html", {"user": request.user})
 
 
 def update(request, id):
+    if not request.user.is_authenticated:
+        return redirect(f"/product/{id}")
+
     product = get_object_or_404(Product, pk=id)
+
     if request.POST:
         name = request.POST.get("product-name")
         price = request.POST.get("price")
@@ -85,10 +109,19 @@ def update(request, id):
             product.desc = desc
             product.save()
             return redirect(f"/product/{id}")
-    return render(request, "update.html", {"product": product})
+
+    return render(request, "update.html", {"user": request.user, "product": product})
 
 
 def delete(request, id):
+    if not request.user.is_authenticated:
+        return redirect(f"/product/{id}")
+
     product = get_object_or_404(Product, pk=id)
     product.delete()
     return redirect("/market")
+
+
+def market(request):
+    products = Product.objects.all()
+    return render(request, "market.html", {"user": request.user, "products": products})
